refactor: make sure consumers import the Draggable Panel from containers
