+++ src/components/containers/DraggablePanel/index.tsx
