+++ src/components/containers/index.ts
@@ -1,3 +1,4 @@
-export { default as Card } from './Card';
-export { default as Modal } from './Modal';
-export { default as ExpandablePanel } from './ExpandablePanel';
+export { default as Card } from "./Card";
+export { default as Modal } from "./Modal";
+export { default as ExpandablePanel } from "./ExpandablePanel";
+export { DraggablePanel } from "./DraggablePanel";
